Add "claim pen" button system
Rather than have the drawer arbitrarily defined by the order
in which clients join, a button is displayed in the UI that will
allow any user to "claim the pen". Once a user has claimed the pen
and is given the status of a drawer, the button is removed from all of
the other users' UIs. If the drawer disconnects, then the button
returns to all of the other users screens, so that they can pick
up the game.