Add new landing page and redirect to app
When users arrive at the main page, they will have the ability
to start their own instance of the game. The app will redirect them
to a URL based on the id of the game as saved to the db.

Get requests to a specific game instance will return a page with
the react app.