Refactor socket.io usage to keep logic at top of comp. hierarchy
As a rule, it is best to keep most of the application logic in a
container component near the top level of the component tree. This
has the benefit of minimizing the number of stateful components in
order to ensure that the application is easier to reason about.

In this particular case, we register all of our socket event
listeners in the top level App component, and then use various
functions to ensure that the correct data is passed down to child
components.

This separation of concerns should allow for easier future
exploration of using systems like Redux to handle changes
to application state in a more sensible way.

Closes #27