Create private socket.io rooms for each game instance
When users get the game app using the unique game instance url,
Express renders the page and passes a single data key for gameId.
The ejs template renders this gameId into a script tag that attaches
the gameId to the window object, for use by the React app.
Subsequently, React mounts the App component and passes in the
gameId as a prop. When the component is done mounting, it emits
a socket.io app:init event that tells the server to add the socket to
a private room based on that gameId. This ensures that events are
only broadcast out to people playing the same game instance.

Closes #17