Conditionally render certain interface components
Depending on the state of application, a user may be in the role
of a drawer or a guesser. The React app keeps track of this state and
sends that state via props to the game interface. Within the interface,
we can use short circuited comparators evaluated within the JSX
to conditionally render components depending on the role of the user.'

Closes #33
Closes #40